Move DefaultAnalyticsRequestExecutor into its own class
