+++ payments-core/src/main/java/com/stripe/android/networking/AnalyticsRequestExecutor.kt
@@ -1,14 +1,6 @@
 package com.stripe.android.networking
 
 import androidx.annotation.RestrictTo
-import androidx.annotation.VisibleForTesting
-import com.stripe.android.Logger
-import com.stripe.android.exception.APIConnectionException
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import java.io.IOException
-import kotlin.coroutines.CoroutineContext
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 fun interface AnalyticsRequestExecutor {
@@ -16,40 +8,4 @@ fun interface AnalyticsRequestExecutor {
      * Execute the fire-and-forget request asynchronously.
      */
     fun executeAsync(request: AnalyticsRequest)
-
-    class Default(
-        private val logger: Logger = Logger.noop(),
-        private val workContext: CoroutineContext = Dispatchers.IO
-    ) : AnalyticsRequestExecutor {
-        private val connectionFactory = ConnectionFactory.Default()
-
-        /**
-         * Make the request and ignore the response
-         *
-         * @return the response status code. Used for testing purposes.
-         */
-        @VisibleForTesting
-        internal fun execute(request: AnalyticsRequest): Int {
-            connectionFactory.create(request).use {
-                try {
-                    // required to trigger the request
-                    return it.responseCode
-                } catch (e: IOException) {
-                    throw APIConnectionException.create(e, request.baseUrl)
-                }
-            }
-        }
-
-        override fun executeAsync(request: AnalyticsRequest) {
-            logger.info("Event: ${request.params[AnalyticsRequestFactory.FIELD_EVENT]}")
-
-            CoroutineScope(workContext).launch {
-                runCatching {
-                    execute(request)
-                }.onFailure {
-                    logger.error("Exception while making analytics request", it)
-                }
-            }
-        }
-    }
 }
