+++ payments-core/src/main/java/com/stripe/android/view/PaymentAuthWebViewActivityViewModel.kt
@@ -14,6 +14,7 @@ import com.stripe.android.networking.AnalyticsEvent
 import com.stripe.android.networking.AnalyticsRequest
 import com.stripe.android.networking.AnalyticsRequestExecutor
 import com.stripe.android.networking.AnalyticsRequestFactory
+import com.stripe.android.networking.DefaultAnalyticsRequestExecutor
 import com.stripe.android.networking.StripeClientUserAgentHeaderFactory
 import com.stripe.android.payments.PaymentFlowResult
 import com.stripe.android.stripe3ds2.init.ui.StripeToolbarCustomization
@@ -121,7 +122,7 @@ internal class PaymentAuthWebViewActivityViewModel(
 
             return PaymentAuthWebViewActivityViewModel(
                 args,
-                AnalyticsRequestExecutor.Default(logger),
+                DefaultAnalyticsRequestExecutor(logger),
                 AnalyticsRequestFactory(
                     application,
                     publishableKey,
